requirements.txt
fastapi
uvicorn[standard]
openai
duckdb
pandas
streamlit
python-dotenv
pytest
altair